@@ -8,8 +8,8 @@
     "@testing-library/user-event": "^13.5.0",
     "history": "^5.0.0",
     "next": "^14.0.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.0.0-beta.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
